bootcamp_daily_progress_fraud_detector/requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
numpy
scikit-learn
websockets
python-dotenv
apscheduler
pytz
pandas
streamlit